# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PHP Admin Dashboard",
        "contact": {
            "email": "Fredowijaya36@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://groupproject-production.up.railway.app/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/contact_us": {
            "get": {
                "tags": [
                    "Contact Us"
                ],
                "summary": "Get all Contact Us datas",
                "operationId": "15d052e3bd72cde4d2f6669e437468dc",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "address": {
                                                "type": "string"
                                            },
                                            "message": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No data found"
                    }
                }
            }
        },
        "/contact_us/add": {
            "post": {
                "tags": [
                    "Contact Us"
                ],
                "summary": "Create a new Contact Us data",
                "operationId": "300d347af98c688a8d3fab220bc2004d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "address",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Contact name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Contact email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "description": "Contact phone",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "address": {
                                        "description": "Contact address",
                                        "type": "string",
                                        "example": "123 Street, City"
                                    },
                                    "message": {
                                        "description": "Contact message",
                                        "type": "string",
                                        "example": "This is a test message"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "phone": {
                                            "type": "string"
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error occurred"
                    }
                }
            }
        },
        "/contact_us/update/{id}": {
            "post": {
                "tags": [
                    "Contact Us"
                ],
                "summary": "Update an Contact Us data",
                "operationId": "3e79b20e4bac27af9752b4fc98efebb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Contact name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Contact email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "description": "Contact phone",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "address": {
                                        "description": "Contact address",
                                        "type": "string",
                                        "example": "123 Street, City"
                                    },
                                    "message": {
                                        "description": "Contact message",
                                        "type": "string",
                                        "example": "This is a test message"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email"
                                                },
                                                "phone": {
                                                    "type": "string"
                                                },
                                                "address": {
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found"
                    },
                    "500": {
                        "description": "Error occurred"
                    }
                }
            }
        },
        "/contact_us/delete/{id}": {
            "delete": {
                "tags": [
                    "Contact Us"
                ],
                "summary": "Delete a contact",
                "operationId": "d63eea2110f2abd9e74cd46aeeb3b561",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contact deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contact not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contact not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}